Finish Contact page responsiveness and email action
