--- src/actions/sendEmail.ts
@@ -1,56 +1,0 @@
-'use server';
-
-import React from 'react';
-import { Resend } from 'resend';
-import { validateString, getErrorMessage } from '@/lib/utils';
-import ContactFormEmail from '@/email/contact-form-email';
-
-const resend = new Resend(process.env.RESEND_API_KEY);
-
-const sendEmail = async (formData: FormData) => {
-  const senderName = formData.get('senderName');
-  const senderEmail = formData.get('senderEmail');
-  const message = formData.get('message');
-
-  // simple server-side validation
-  if (!validateString(senderName, 500)) {
-    return {
-      error: 'Invalid sender name',
-    };
-  }
-  if (!validateString(senderEmail, 500)) {
-    return {
-      error: 'Invalid sender email',
-    };
-  }
-  if (!validateString(message, 5000)) {
-    return {
-      error: 'Invalid message',
-    };
-  }
-
-  let data;
-  try {
-    data = await resend.emails.send({
-      from: 'My Portfolio Contact Form <onboarding@resend.dev>',
-      to: 'ganzhiheng1@gmail.com',
-      subject: senderName + ' left a message',
-      reply_to: senderEmail,
-      react: React.createElement(ContactFormEmail, {
-        message: message,
-        senderName: senderName,
-        senderEmail: senderEmail,
-      }),
-    });
-  } catch (error: unknown) {
-    return {
-      error: getErrorMessage(error),
-    };
-  }
-
-  return {
-    data,
-  };
-};
-
-export { sendEmail };
