+++ src/data/errors.ts
@@ -1,0 +1,4 @@
+export const serverError = {
+  error: 'Something went wrong. Please try again.',
+  message: 'Internal Server Error',
+};
